+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [Unreleased]
+
+### Added
+- Implement simulate command line utility (from @justjasongreen)
+
+
 ## [1.0.0a4] - 2016-07-29
 
 ### Fixed
@@ -50,6 +56,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Set up project (from @justjasongreen)
 
 
+[Unreleased]: https://github.com/justjasongreen/predictive_punter/compare/1.0.0a4...HEAD
 [1.0.0a4]: https://github.com/justjasongreen/predictive_punter/compare/1.0.0a3...1.0.0a4
 [1.0.0a3]: https://github.com/justjasongreen/predictive_punter/compare/1.0.0a2...1.0.0a3
 [1.0.0a2]: https://github.com/justjasongreen/predictive_punter/compare/1.0.0a1...1.0.0a2